functions/publish_ticket/requirements.txt
functions-framework==3.5.0
google-cloud-pubsub==2.21.1
